Improve static handlers.
The staticFile handler will now correctly handle gzipped assets by setting the content-encoding
header if it is a known content type, or setting `content-type: application/gzip` if not. Also
enumerates a basic set of known content types, which can easily be extended in the future. Adds a
recursive addStaticDir method which adds everything inside a directory as a static route.